add email controller and route
