+++ src/express.js
@@ -6,8 +6,7 @@ import teacherRoutes from './routes/teacher.route.js';
 import supplyRoutes from './routes/supply.route.js';
 import studentRoutes from './routes/student.route.js';
 import donationRoutes from './routes/donation.route.js';
-import nodemailer from 'nodemailer';
-import smtpTransport from 'nodemailer-smtp-transport';
+import emailRoutes from './routes/email.route.js';
 
 const app = express();
 
@@ -31,6 +30,7 @@ app.use('/', teacherRoutes);
 app.use('/', supplyRoutes);
 app.use('/', studentRoutes);
 app.use('/', donationRoutes);
+app.use('/', emailRoutes);
 
 app.get('/', (req, res) =>
     res.status(200).send('Donation Web API is running on Render')
@@ -39,33 +39,6 @@ app.get('/', (req, res) =>
 app.get('/ping', (req, res) => {
     const serverUp = { status: true };
     res.status(200).json(serverUp);
-});
-
-app.post('/sendEmail', (req, res) => {
-    console.log('in sendEmail');
-    let transporter = nodemailer.createTransport(smtpTransport({
-        service: 'gmail',
-        host: 'smtp.gmail.com',
-        auth: {
-            user: 'tsdcapstone@gmail.com',
-            pass: 'frlnvucfltrtkavq'
-        }
-    }));
-    let mailOptions = {
-        from: 'tsdcapstone@gmail.com',
-        to: 'alicefisher100@gmail.com',
-        subject: 'testing',
-        text: 'test email'
-    };
-
-    transporter.sendMail(mailOptions, (error, info) => {
-        if (error) {
-            console.log(error);
-        } else {
-            console.log('Email sent: ' + info.response);
-            res.sendStatus(200);
-        }
-    })
 });
 
 export default app;
