+++ src/routes/email.route.js
@@ -1,0 +1,8 @@
+import express from 'express';
+import emailCtrl from '../controllers/email.controller.js';
+
+const router = express.Router();
+
+router.route('/email').post(emailCtrl.send);
+
+export default router;
